extract session class
